requirements.txt
Django==1.3.1
django-tastypie==0.9.10
django-uuidfield==0.2
mimeparse==0.1.3
netaddr==0.7.6
python-dateutil==1.5
python-digest==1.7
uuid==1.30
wsgiref==0.1.2
